Migrate to Xunit + code clean up.
