+++ Source/Bogus.Tests/Bogus.Tests.csproj
@@ -43,10 +43,6 @@
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\packages\Newtonsoft.Json\lib\net45\Newtonsoft.Json.dll</HintPath>
     </Reference>
-    <Reference Include="nunit.framework, Version=3.6.1.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
-      <HintPath>..\packages\NUnit.3.6.1\lib\net45\nunit.framework.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="Rant, Version=3.0.508.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Rant.3.0.508\lib\net45\Rant.dll</HintPath>
     </Reference>
@@ -59,20 +55,32 @@
     <Reference Include="Microsoft.CSharp" />
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
+    <Reference Include="xunit.abstractions, Version=2.0.0.0, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c, processorArchitecture=MSIL">
+      <HintPath>..\packages\xunit.abstractions.2.0.1\lib\net35\xunit.abstractions.dll</HintPath>
+    </Reference>
+    <Reference Include="xunit.assert, Version=2.2.0.3545, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c, processorArchitecture=MSIL">
+      <HintPath>..\packages\xunit.assert.2.2.0\lib\netstandard1.1\xunit.assert.dll</HintPath>
+    </Reference>
+    <Reference Include="xunit.core, Version=2.2.0.3545, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c, processorArchitecture=MSIL">
+      <HintPath>..\packages\xunit.extensibility.core.2.2.0\lib\netstandard1.1\xunit.core.dll</HintPath>
+    </Reference>
+    <Reference Include="xunit.execution.desktop, Version=2.2.0.3545, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c, processorArchitecture=MSIL">
+      <HintPath>..\packages\xunit.extensibility.execution.2.2.0\lib\net452\xunit.execution.desktop.dll</HintPath>
+    </Reference>
     <Reference Include="Z.ExtensionMethods.WithTwoNamespace, Version=2.0.10.0, Culture=neutral, PublicKeyToken=59b66d028979105b, processorArchitecture=MSIL">
       <HintPath>..\packages\Z.ExtensionMethods.WithTwoNamespace.2.0.10\lib\net45\Z.ExtensionMethods.WithTwoNamespace.dll</HintPath>
       <Private>True</Private>
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AddressTest.cs" />
-    <Compile Include="CardTests.cs" />
-    <Compile Include="CommerceTest.cs" />
-    <Compile Include="CompanyTest.cs" />
-    <Compile Include="DatabaseTests.cs" />
-    <Compile Include="DateTest.cs" />
+    <Compile Include="DataSetTests\AddressTest.cs" />
+    <Compile Include="DataSetTests\CardTests.cs" />
+    <Compile Include="DataSetTests\CommerceTest.cs" />
+    <Compile Include="DataSetTests\CompanyTest.cs" />
+    <Compile Include="DataSetTests\DatabaseTests.cs" />
+    <Compile Include="DataSetTests\DateTest.cs" />
     <Compile Include="ExtensionsForTesting.cs" />
-    <Compile Include="FinanceTest.cs" />
+    <Compile Include="DataSetTests\FinanceTest.cs" />
     <Compile Include="FluentTests.cs" />
     <Compile Include="GitHubIssues\Issue10.cs" />
     <Compile Include="GitHubIssues\Issue13.cs" />
@@ -83,23 +91,23 @@
     <Compile Include="GitHubIssues\Issue55.cs" />
     <Compile Include="GitHubIssues\Issue61.cs" />
     <Compile Include="GitHubIssues\Issue70.cs" />
-    <Compile Include="HackerTests.cs" />
+    <Compile Include="DataSetTests\HackerTests.cs" />
     <Compile Include="HandleBarTests.cs" />
-    <Compile Include="ImageTest.cs" />
-    <Compile Include="InternetTests.cs" />
-    <Compile Include="LoremTests.cs" />
-    <Compile Include="NameTests.cs" />
+    <Compile Include="DataSetTests\ImageTest.cs" />
+    <Compile Include="DataSetTests\InternetTests.cs" />
+    <Compile Include="DataSetTests\LoremTests.cs" />
+    <Compile Include="DataSetTests\NameTests.cs" />
     <Compile Include="PersonTest.cs" />
-    <Compile Include="PhoneNumbersTest.cs" />
+    <Compile Include="DataSetTests\PhoneNumbersTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RandomizerTest.cs" />
-    <Compile Include="RantTests.cs" />
+    <Compile Include="DataSetTests\RantTests.cs" />
     <Compile Include="Rant_Generator.cs" />
     <Compile Include="README_Generator.cs" />
     <Compile Include="RuleSetTests.cs" />
     <Compile Include="SeededTest.cs" />
     <Compile Include="StrictModeTests.cs" />
-    <Compile Include="SystemTest.cs" />
+    <Compile Include="DataSetTests\SystemTest.cs" />
     <Compile Include="UniquenessTests.cs" />
     <Compile Include="ValidationDeltaTests.cs" />
   </ItemGroup>
@@ -118,6 +126,7 @@
       <Name>Bogus</Name>
     </ProjectReference>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>
