+++ Source/Bogus.Tests/DataSetTests/ImageTest.cs
@@ -1,24 +1,20 @@
 using System.Net;
 using Bogus.DataSets;
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 
-namespace Bogus.Tests
+namespace Bogus.Tests.DataSetTests
 {
-    [TestFixture]
     public class ImageTest : SeededTest
     {
-        private Images image;
-
-        [SetUp]
-        public void BeforeEachTest()
+        public ImageTest()
         {
             image = new Images();
         }
 
+        private readonly Images image;
 
-        [Test]
-        [Explicit]
+        [Fact(Skip = "Explicit")]
         public void DownloadAllTest()
         {
             var wc = new WebClient();
@@ -37,20 +33,20 @@ namespace Bogus.Tests
             wc.DownloadFile(image.Transport(), "transport.jpg");
         }
 
-        [Test]
-        public void url_generated_should_have_https()
-        {
-            image.Sports(https: true).Should().StartWith("https://");
-        }
-
-        [Test]
+        [Fact]
         public void svg_data_url()
         {
             image.DataUri(200, 300).Dump();
 
             image.DataUri(200, 300).Should()
-                .Be("data:image/svg+xml;charset=UTF-8,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20version%3D%221.1%22%20baseProfile%3D%22full%22%20width%3D%22200%22%20height%3D%22300%22%3E%20%3Crect%20width%3D%22100%25%22%20height%3D%22100%25%22%20fill%3D%22grey%22%2F%3E%20%20%3Ctext%20x%3D%220%22%20y%3D%2220%22%20font-size%3D%2220%22%20text-anchor%3D%22start%22%20fill%3D%22white%22%3E200x300%3C%2Ftext%3E%20%3C%2Fsvg%3E");
+                .Be(
+                    "data:image/svg+xml;charset=UTF-8,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20version%3D%221.1%22%20baseProfile%3D%22full%22%20width%3D%22200%22%20height%3D%22300%22%3E%20%3Crect%20width%3D%22100%25%22%20height%3D%22100%25%22%20fill%3D%22grey%22%2F%3E%20%20%3Ctext%20x%3D%220%22%20y%3D%2220%22%20font-size%3D%2220%22%20text-anchor%3D%22start%22%20fill%3D%22white%22%3E200x300%3C%2Ftext%3E%20%3C%2Fsvg%3E");
         }
-   
+
+        [Fact]
+        public void url_generated_should_have_https()
+        {
+            image.Sports(https: true).Should().StartWith("https://");
+        }
     }
 }
