+++ Source/Bogus.Tests/DataSetTests/LoremTests.cs
@@ -1,28 +1,24 @@
-using System;
-using System.Linq;
 using Bogus.DataSets;
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 using Z.ExtensionMethods;
 
-namespace Bogus.Tests
+namespace Bogus.Tests.DataSetTests
 {
-    [TestFixture]
     public class LoremTests : SeededTest
     {
-        private Lorem lorem;
-
-        [SetUp]
-        public void BeforeEachTest()
+        public LoremTests()
         {
             lorem = new Lorem();
         }
 
-        [Test]
+        private readonly Lorem lorem;
+
+        [Fact]
         public void can_get_3_words()
         {
             var test = lorem.Words();
-            
+
             test.Dump();
 
             test.Should()
@@ -33,7 +29,7 @@ namespace Bogus.Tests
             //test.ForEach(w => w.Should().NotBeNullOrWhiteSpace());
         }
 
-        [Test]
+        [Fact]
         public void can_get_5_words()
         {
             var test = lorem.Words(5);
@@ -52,39 +48,47 @@ namespace Bogus.Tests
             //test.ForEach(w => w.Should().NotBeNullOrWhiteSpace());
         }
 
-        [Test]
-        public void can_get_a_sentence()
-        {
-            lorem.Sentence().Split(' ').Length.Should().BeGreaterThan(3);
-        }
-
-        [Test]
-        public void can_get_a_sentence_with_options()
-        {
-            lorem.Sentence(5).Split(' ').Length.Should().Be(5);
-        }
-
-        [Test]
-        public void can_get_a_setnance_with_range_option()
-        {
-            lorem.Sentence(10, 5).Split(' ').Length.Should().Be(13);
-        }
-
-        [Test]
+        [Fact]
         public void can_get_a_paragraph()
         {
             lorem.Paragraph()
                 .Split(". ").Length.Should().Be(5); // para of 5 sentances.
         }
 
-        [Test]
+        [Fact]
+        public void can_get_a_random_word()
+        {
+            lorem.Word().Should().Be("id");
+            lorem.Word().Should().Be("aut");
+            lorem.Word().Should().Be("vel");
+        }
+
+        [Fact]
+        public void can_get_a_sentence()
+        {
+            lorem.Sentence().Split(' ').Length.Should().BeGreaterThan(3);
+        }
+
+        [Fact]
+        public void can_get_a_sentence_with_options()
+        {
+            lorem.Sentence(5).Split(' ').Length.Should().Be(5);
+        }
+
+        [Fact]
+        public void can_get_a_setnance_with_range_option()
+        {
+            lorem.Sentence(10, 5).Split(' ').Length.Should().Be(13);
+        }
+
+        [Fact]
         public void can_get_paragraphs()
         {
             lorem.Paragraphs()
                 .Split("\n\n").Length.Should().Be(3);
         }
 
-        [Test]
+        [Fact]
         public void can_get_some_letters()
         {
             var c = lorem.Letter();
@@ -96,28 +100,19 @@ namespace Bogus.Tests
             chars.Should().Be("eiblrueeulrtiorismecntonniaeaaumrumclrquoqaeoiehdtueuteisquagsieuiuturutunuuaiuamisseqvnqeratepilptt");
         }
 
-        [Test]
-        public void can_get_a_random_word()
-        {
-            lorem.Word().Should().Be("id");
-            lorem.Word().Should().Be("aut");
-            lorem.Word().Should().Be("vel");
-
-        }
-
-        [Test]
-        public void can_get_some_lorem_text()
-        {
-            lorem.Text().Split(" ").Length.Should().BeGreaterThan(5);
-        }
-
-        [Test]
+        [Fact]
         public void can_get_some_lorem_lines()
         {
             lorem.Lines().Split(" ").Length.Should().BeGreaterThan(5);
         }
 
-        [Test]
+        [Fact]
+        public void can_get_some_lorem_text()
+        {
+            lorem.Text().Split(" ").Length.Should().BeGreaterThan(5);
+        }
+
+        [Fact]
         public void can_slugify_lorem()
         {
             lorem.Slug(5).Should().Be("id-aut-vel-facilis-aperiam");
