+++ Source/Bogus.Tests/DataSetTests/PhoneNumbersTest.cs
@@ -1,35 +1,33 @@
 using Bogus.DataSets;
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 
-namespace Bogus.Tests
+namespace Bogus.Tests.DataSetTests
 {
-    [TestFixture]
-    public class PhoneNumbersTest: SeededTest
+    public class PhoneNumbersTest : SeededTest
     {
-        private PhoneNumbers phone;
-
-        [SetUp]
-        public void BeforeEachTest()
+        public PhoneNumbersTest()
         {
             phone = new PhoneNumbers();
         }
 
-        [Test]
+        private readonly PhoneNumbers phone;
+
+        [Fact]
         public void can_get_phone_number()
         {
             phone.PhoneNumber()
                 .Should().Be("186-060-6439 x1750");
         }
 
-        [Test]
+        [Fact]
         public void can_get_phone_number_of_specific_format()
         {
             phone.PhoneNumber("## ### ####")
                 .Should().Be("61 860 6064");
         }
 
-        [Test]
+        [Fact]
         public void can_get_phone_number_via_formats_index()
         {
             phone.PhoneNumberFormat(1)
