+++ Source/Bogus.Tests/HandleBarTests.cs
@@ -1,51 +1,15 @@
-﻿using System;
-using System.Linq;
-using System.Reflection;
-using System.Text.RegularExpressions;
-using Bogus.DataSets;
-using NUnit.Framework;
+﻿using Xunit;
 
 namespace Bogus.Tests
 {
-    [TestFixture]
     public class HandleBarTests
     {
-        [OneTimeSetUp]
-        public void BeforeRunningTestSession()
-        {
-
-        }
-
-        [OneTimeTearDown]
-        public void AfterRunningTestSession()
-        {
-
-        }
-
-
-        [SetUp]
-        public void BeforeEachTest()
-        {
-
-        }
-
-        [TearDown]
-        public void AfterEachTest()
-        {
-
-        }
-
-
-
-    
-        [Test]
+        [Fact]
         public void parse_test()
         {
             var f = new Faker();
             Tokenizer.Parse("{{name.lastName}}, {{name.firstName}} {{name.suffix}}", f.Name)
                 .Dump();
-
         }
     }
-
 }
