+++ Source/Bogus.Tests/RandomizerTest.cs
@@ -1,19 +1,15 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Runtime.InteropServices;
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 
 namespace Bogus.Tests
 {
-    [TestFixture]
     public class RandomizerTest : SeededTest
     {
         private Randomizer r;
 
-        [OneTimeSetUp]
-        public void BeforeRunningTestSession()
+        public RandomizerTest()
         {
             r = new Randomizer();
         }
@@ -21,28 +17,31 @@ namespace Bogus.Tests
         public enum Foo
         {
             ExcludeMe,
-            A,B,C,D
+            A,
+            B,
+            C,
+            D
         }
 
-        [Test]
+        [Fact]
         public void pick_an_enum()
         {
             var f = r.Enum<Foo>();
             f.Should().Be(Foo.C);
         }
 
-        [Test]
+        [Fact]
         public void exclude_an_enum()
         {
             //seeded value of 14 gets "ExcludeMe", ensure exclude works.
             Randomizer.Seed = new Random(14);
-            var f = r.Enum( exclude: Foo.ExcludeMe);
+            var f = r.Enum(exclude: Foo.ExcludeMe);
             f.ToString().Dump();
 
             f.Should().NotBe(Foo.ExcludeMe);
         }
 
-        [Test]
+        [Fact]
         public void exclude_all_throws_an_error()
         {
             Action act = () => r.Enum(Foo.ExcludeMe, Foo.A, Foo.B, Foo.C, Foo.D);
@@ -50,14 +49,14 @@ namespace Bogus.Tests
             act.ShouldThrow<ArgumentException>();
         }
 
-        [Test]
+        [Fact]
         public void can_replace_numbers_or_letters_using_asterisk()
         {
             r.Replace("***")
                 .Should().Be("CQ6");
         }
 
-        [Test]
+        [Fact]
         public void can_get_random_word()
         {
             r.Word().Should().Be("Court");
@@ -65,7 +64,7 @@ namespace Bogus.Tests
             r.Word().Should().Be("Movies & Clothing");
         }
 
-        [Test]
+        [Fact]
         public void can_get_some_random_words()
         {
             r.Words().Should().Be("Soft deposit");
@@ -73,83 +72,82 @@ namespace Bogus.Tests
             r.Words().Should().Be("Corner Handcrafted Frozen Chair transmitting");
         }
 
-        [Test]
+        [Fact]
         public void can_shuffle_some_enumerable()
         {
             new string(r.Shuffle("123456789").ToArray())
                 .Should().Be("628753491");
         }
 
-        [Test]
+        [Fact]
         public void can_get_random_locale()
         {
             r.RandomLocale().Should().Be("ko");
         }
 
-        [Test]
+        [Fact]
         public void exclusive_int_maxvalue_number()
         {
             var max = r.Number(int.MaxValue - 1, int.MaxValue);
             max.Should().Be(int.MaxValue - 1);
         }
 
-        [Test]
+        [Fact]
         public void random_bool()
         {
             r.Bool().Should().BeFalse();
         }
 
-        [Test]
+        [Fact]
         public void can_get_some_alpha_chars()
         {
             r.AlphaNumeric(20).Should().Be("l3tn1m1ohax6ql31pw1u");
         }
 
-        [Test]
+        [Fact]
         public void generate_double_with_min_and_max()
         {
             r.Double(2.5, 2.9).Should().BeInRange(2.74140891332244, 2.74140891332246);
         }
 
-        [Test]
+        [Fact]
         public void generate_decimal_with_min_and_max()
         {
             r.Decimal(2.2m, 5.2m).Should().Be(4.0105668499183690m);
         }
 
-        [Test]
+        [Fact]
         public void generate_float_with_min_and_max()
         {
             r.Float(2.7f, 3.9f).Should().BeInRange(3.424226f, 3.424228f);
-
         }
 
-        [Test]
+        [Fact]
         public void generate_byte()
         {
             r.Byte(1, 128).Should().Be(78);
         }
 
-        [Test]
+        [Fact]
         public void generate_some_bytes()
         {
             r.Bytes(20).Should()
                 .Equal(218, 35, 156, 76, 224, 196, 45, 215, 227, 196, 168, 150, 23, 242, 85, 178, 101, 200, 89, 189);
         }
 
-        [Test]
+        [Fact]
         public void generate__sbyte()
         {
-            r.SByte(max:0).Should().Be(-51);
+            r.SByte(max: 0).Should().Be(-51);
         }
 
-        [Test]
+        [Fact]
         public void generate_uint32()
         {
             r.UInt(99, 200).Should().Be(160);
         }
 
-        [Test]
+        [Fact]
         public void generate_unit32_many()
         {
             r.UInt().Should().Be(2592108469u);
@@ -158,13 +156,13 @@ namespace Bogus.Tests
             r.UInt().Should().Be(2775978649u);
         }
 
-        [Test]
+        [Fact]
         public void generate_int32()
         {
             r.Int(max: 0).Should().Be(-425714706);
         }
 
-        [Test]
+        [Fact]
         public void generate_int32_many()
         {
             r.Int().Should().Be(1296054233);
@@ -173,13 +171,13 @@ namespace Bogus.Tests
             r.Int().Should().Be(108870444);
         }
 
-        [Test]
+        [Fact]
         public void generate_uint64()
         {
             r.ULong(99, 9999).Should().Be(6074);
         }
 
-        [Test]
+        [Fact]
         public void generate_uint64_many()
         {
             r.ULong().Should().Be(11133021102928879616UL);
@@ -188,13 +186,13 @@ namespace Bogus.Tests
             r.ULong().Should().Be(11922737513106253824UL);
         }
 
-        [Test]
+        [Fact]
         public void generate_int64()
         {
             r.Long(max: 0).Should().Be(-3656861485390335055L);
         }
 
-        [Test]
+        [Fact]
         public void generate_int64_many()
         {
             r.Long().Should().Be(1909649066074105698L);
@@ -204,13 +202,13 @@ namespace Bogus.Tests
             r.Long().Should().Be(-8566699986853958425L);
         }
 
-        [Test]
+        [Fact]
         public void generate_int16()
         {
             r.Short(max: 0).Should().Be(-12992);
         }
 
-        [Test]
+        [Fact]
         public void generate_int16_many()
         {
             r.Short().Should().Be(6784);
@@ -219,19 +217,19 @@ namespace Bogus.Tests
             r.Short().Should().Be(9589);
         }
 
-        [Test]
+        [Fact]
         public void generate_uint16()
         {
             r.UShort().Should().Be(39552);
         }
 
-        [Test]
+        [Fact]
         public void generate_char()
         {
             r.Char().Should().Be('\u9a80');
         }
 
-        [Test]
+        [Fact]
         public void generate_some_chars()
         {
             r.Chars(count: 10).Should().Equal(
@@ -247,18 +245,14 @@ namespace Bogus.Tests
                 '\u4d58');
         }
 
-        [Test]
+        [Fact]
         public void random_word_tests()
         {
             //r.Words(3).Should().Be("");
             //r.Words(5).Split(' ').Length.Should().Be(4);
             r.WordsArray(5).Length.Should().Be(5);
-            r.WordsArray(1, 80).Length.Should().BeInRange(1,80); //.Should().BeInRange(1, 80);
+            r.WordsArray(1, 80).Length.Should().BeInRange(1, 80); //.Should().BeInRange(1, 80);
             r.WordsArray(10, 20).Length.Should().BeInRange(10, 20);
-
         }
     }
-
-
-
 }
