+++ Source/Bogus.Tests/SeededTest.cs
@@ -1,11 +1,11 @@
-﻿using NUnit.Framework;
+﻿using Xunit;
 
 namespace Bogus.Tests
 {
+    [Collection("Seeded Test")]
     public class SeededTest
-    { 
-        [SetUp]
-        public void BeforeEachTest()
+    {
+        public SeededTest()
         {
             //set the random gen manually to a seeded value
             Randomizer.Seed = new System.Random(3116);
