+++ Source/Bogus.Tests/ValidationDeltaTests.cs
@@ -1,13 +1,11 @@
-﻿using System;
-using FluentAssertions;
-using NUnit.Framework;
+﻿using FluentAssertions;
+using Xunit;
 
 namespace Bogus.Tests
 {
-    [TestFixture]
     public class ValidationDeltaTests
     {
-        [Test]
+        [Fact]
         public void should_be_valid_and_not_throw_exception_on_complete_rules_with_ignore()
         {
             var testOrders = new Faker<Examples.Order>()
@@ -16,11 +14,11 @@ namespace Bogus.Tests
                 .RuleFor(o => o.OrderId, f => 3343)
                 .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5));
             var result = testOrders.Validate();
-            Assert.DoesNotThrow(() => testOrders.AssertConfigurationIsValid());
+            testOrders.AssertConfigurationIsValid();
             result.ShouldBeEquivalentTo(true);
         }
 
-        [Test]
+        [Fact]
         public void should_not_be_valid_and_throw_exception_on_incomplete_rules()
         {
             var testOrders = new Faker<Examples.Order>()
@@ -31,7 +29,7 @@ namespace Bogus.Tests
             result.ShouldBeEquivalentTo(false);
         }
 
-        [Test]
+        [Fact]
         public void should_throw_exception_on_incomplete_rules()
         {
             var testOrders = new Faker<Examples.Order>()
@@ -40,42 +38,42 @@ namespace Bogus.Tests
             Assert.Throws<ValidationException>(() => testOrders.AssertConfigurationIsValid());
         }
 
-        [Test]
+        [Fact]
         public void should_be_valid_and_no_excpetions_on_complete_rules()
         {
             var testOrders = new Faker<Examples.Order>()
-                .StrictMode(true)
-                .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5))
-                .RuleFor(o => o.OrderId, f => f.Random.Number(2, 5))
-                .RuleFor(o => o.Item, f => f.Lorem.Sentence())
+                    .StrictMode(true)
+                    .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5))
+                    .RuleFor(o => o.OrderId, f => f.Random.Number(2, 5))
+                    .RuleFor(o => o.Item, f => f.Lorem.Sentence())
                 ;
             var result = testOrders.Validate();
-            Assert.DoesNotThrow(() => testOrders.AssertConfigurationIsValid());
+            testOrders.AssertConfigurationIsValid();
             result.ShouldBeEquivalentTo(true);
         }
 
-        [Test]
+        [Fact]
         public void should_be_valid_no_excpetions_on_incomplete_rules_when_strict_false()
         {
             var testOrders = new Faker<Examples.Order>()
                 .StrictMode(false)
                 .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5));
             var result = testOrders.Validate();
-            Assert.DoesNotThrow(() => testOrders.AssertConfigurationIsValid());
+            testOrders.AssertConfigurationIsValid();
             result.ShouldBeEquivalentTo(true);
         }
 
-        [Test]
+        [Fact]
         public void should_be_valid_on_complete_rules()
         {
             var testOrders = new Faker<Examples.Order>()
-                .StrictMode(true)
-                .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5))
-                .RuleFor(o => o.OrderId, f => f.Random.Number(2, 5))
-                .RuleFor(o => o.Item, f => f.Lorem.Sentence())
+                    .StrictMode(true)
+                    .RuleFor(o => o.Quantity, f => f.Random.Number(2, 5))
+                    .RuleFor(o => o.OrderId, f => f.Random.Number(2, 5))
+                    .RuleFor(o => o.Item, f => f.Lorem.Sentence())
                 ;
             var result = testOrders.Validate();
-            Assert.DoesNotThrow(() => testOrders.AssertConfigurationIsValid());
+            testOrders.AssertConfigurationIsValid();
             result.ShouldBeEquivalentTo(true);
         }
     }
