+++ Source/Bogus.Tests/app.config
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <configuration>
-  
-  
-<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2"/></startup></configuration>
+
+
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.2" />
+  </startup>
+</configuration>
