+++ Source/Bogus.Tests/packages.config
@@ -1,8 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
+
 <packages>
   <package id="FluentAssertions" version="4.19.2" targetFramework="net451" />
-  <package id="NUnit" version="3.6.1" targetFramework="net451" />
   <package id="NUnit.ConsoleRunner" version="3.6.1" targetFramework="net451" />
   <package id="Rant" version="3.0.508" targetFramework="net462" />
+  <package id="xunit" version="2.2.0" targetFramework="net462" />
+  <package id="xunit.abstractions" version="2.0.1" targetFramework="net462" />
+  <package id="xunit.assert" version="2.2.0" targetFramework="net462" />
+  <package id="xunit.core" version="2.2.0" targetFramework="net462" />
+  <package id="xunit.extensibility.core" version="2.2.0" targetFramework="net462" />
+  <package id="xunit.extensibility.execution" version="2.2.0" targetFramework="net462" />
   <package id="Z.ExtensionMethods.WithTwoNamespace" version="2.0.10" targetFramework="net451" />
 </packages>
