+++ Source/paket.dependencies
@@ -9,11 +9,11 @@ group build
     
     source https://www.nuget.org/api/v2
 
-    nuget FAKE == 4.57.4
+    nuget FAKE == 4.61.2
     nuget FSharp.Data
     nuget Z.ExtensionMethods.WithNamespace
     nuget DotNetZip == 1.10.1
     nuget NuGet.CommandLine == 3.5.0
-    nuget NUnit.ConsoleRunner == 3.6.1
+    nuget xunit.runner.console == 2.2.0
     nuget Newtonsoft.Json
     nuget secure-file
